requirements.txt
aiokafka==0.8.0
alembic==1.10.4
asyncpg==0.27.0
asyncssh==2.13.1
fastapi[all]==0.95.1
pydantic==1.10.7
PyYAML==6.0
SQLAlchemy==2.0.12
trafaret==2.1.1
uvicorn==0.22.0
websockets==11.0.2
yarl==1.9.2
typer~=0.9.0
aiohttp~=3.8.4
python-jose==3.3.0